improve the GetMerklePathByTransactionId
